+++ tests/EvmLookupServer/CMakeLists.txt
@@ -20,19 +20,3 @@ target_link_libraries(Test_EvmLookupServer
 )
 
 add_test(NAME Test_EvmLookupServer COMMAND Test_EvmLookupServer)
-
-
-# evm accounts tests
-add_executable(Test_EvmAccounts Test_EvmAccounts.cpp)
-
-target_include_directories(Test_EvmAccounts
-                            PUBLIC
-                            ${CMAKE_SOURCE_DIR}/src
-)
-
-target_link_libraries(Test_EvmAccounts
-                      PUBLIC
-                      AccountData
-)
-
-add_test(NAME Test_EvmAccounts COMMAND Test_EvmAccounts)
